steps: Move some code to its proper place
Mostly moves functions to a better place in the module hierarchy.
Notably for lone functions in a separate "common" module, which are
specific and only called from one place, move them to the module that
uses them.
