+++ dnf-behave-tests/features/steps/common/__init__.py
@@ -1,9 +1,6 @@
 from .behave_ext import check_context_table
-from .behave_ext import parse_context_table
 from .cmd import run
 from .cmd import run_in_context
-from .cmd import get_boot_time
-from .cmd import start_server_based_on_type
 from .checksum import sha256_checksum
 from .dnf import parse_history_info
 from .dnf import parse_history_list
@@ -19,11 +16,6 @@ from .file import ensure_file_exists
 from .file import delete_directory
 from .file import delete_file
 from .file import file_timestamp
-from .module import get_modules_state
-from .repo import generate_metalink
 from .rpm import RPM
 from .rpm import diff_rpm_lists
 from .rpmdb import get_rpmdb_rpms
-from .shell import stdout_from_shell
-from .string import extract_section_content_from_text
-from .string import lines_match_to_regexps_line_by_line
