+++ dnf-behave-tests/features/steps/repo.py
@@ -9,8 +9,10 @@ import os
 import parse
 
 from common import *
+from common.checksum import sha256_checksum
 from common.rpmdb import get_rpmdb_rpms
 from environment import osrelease
+from fixtures import start_server_based_on_type
 
 
 def repo_config(repo, new={}):
@@ -80,6 +82,36 @@ def generate_repodata(context, repo):
         run_in_context(context, "modifyrepo_c --mdtype=modules %s %s" % (modules_filename, repodata_path))
 
     context.repos[repo] = True
+
+
+def generate_metalink(destdir, url):
+    metalink = """<?xml version="1.0" encoding="utf-8"?>
+<metalink version="3.0" xmlns="http://www.metalinker.org/" xmlns:mm0="http://fedorahosted.org/mirrormanager">
+  <files>
+    <file name="repomd.xml">
+      <mm0:timestamp>1550000000</mm0:timestamp>
+      <size>{size}</size>
+      <verification>
+        <hash type="sha256">{csum}</hash>
+      </verification>
+      <resources>
+        <url protocol="{schema}" type="{schema}">{url}/repodata/repomd.xml</url>
+      </resources>
+    </file>
+  </files>
+</metalink>
+"""
+    schema = url.split(':')[0]
+    with open(os.path.join(destdir, 'repodata', 'repomd.xml')) as f:
+        repomd = f.read()
+    with open(os.path.join(destdir, 'metalink.xml'), 'w') as f:
+        data = metalink.format(
+            size=len(repomd),
+            csum=sha256_checksum(repomd.encode('utf-8')),
+            schema=schema,
+            url=url,
+        )
+        f.write(data + '\n')
 
 
 @behave.step("I use repository \"{repo}\"")
